full project not oop

+++ actions/tickets/remove-all.php
@@ -12,8 +12,7 @@ $query = $db->prepare("SELECT * FROM `users` WHERE `id` = :id");
 $query->execute(['id' => $_SESSION['user']]);
 $user = $query->fetch(PDO::FETCH_ASSOC);
 
-
-$query = $db->prepare("DELETE FROM `tasks`");
-$query->execute();
+$query = $db->prepare("DELETE FROM `tasks` WHERE `user_id` = :user_id");
+$query->execute(['user_id' => $user['id']]);
 
 header('Location: /тестовое/my-tickets.php');
